support host, object instance  export id  issue #599
